Reordered Entity package and added a Utils package where i have put the HibernateUtil, ProductDao for  DB actions UserDao for DB actions UserVerifier for checking user existence and type of user
