+++ src/main/java/sample/Utils/UserVerifier.java
@@ -1,4 +1,4 @@
-package sample;
+package sample.Utils;
 
 import org.hibernate.Session;
 import org.hibernate.query.Query;
@@ -6,23 +6,22 @@ import sample.entity.User;
 
 public class UserVerifier {
 
-    private static Session session = HibernateUtil.getSession();
+    private Session session = HibernateUtil.getSession();
 
     public boolean userExist(User userToVerify) {
         Query<User> query = session.createQuery(
-                "select u from User u where username = :username and  password = :password and userType = 'employee'", User.class);
+                "select u from User u where username = :username and  password = :password", User.class);
         query.setParameter("username", userToVerify.getUsername());
         query.setParameter("password", userToVerify.getPassword());
         return query.getResultList().size() > 0;
     }
 
-    public static boolean isUserCustomerType(User userToVerify) {
+    public boolean isUserEmployeeType(User userToVerify) {
         Query<User> query = session.createQuery(
-                "select u from User u where username = :username and  password = :password and userType = 'customer'", User.class);
+                "select u from User u where username = :username and  password = :password", User.class);
         query.setParameter("username", userToVerify.getUsername());
         query.setParameter("password", userToVerify.getPassword());
-        //return query.getResultList().equals("customer");
-       //return query.getSingleResult().getUserType().equals("customer");
-        return query.getResultList().size() > 0;
+        return "employee".equals(query.getSingleResult().getUserType());
     }
+
 }
